+++ CHANGES.md
@@ -3,7 +3,7 @@ ringpop-go changes
 
 Next Release
 ------------
-* Feature: testpop takes an additional -stats flag that can be used to emit
+* Feature: testpop takes an additional -stats-udp or -stats-file flag that can be used to emit
   statsd messages to a file or over the network. #144
 
 v0.5.0